Extract parsing of user commands into a class
