+++ README.md
@@ -1,4 +1,4 @@
-# Spark
-Spark is a chat-bot that helps you keep track of your tasks!
+# spark.spark
+spark.spark is a chat-bot that helps you keep track of your tasks!
 
 Made for CS2103T Individual Project (iP).
