--- src/main/java/SparkExceptions/FileExceptions/FileException.java
@@ -1,9 +1,0 @@
-package SparkExceptions.FileExceptions;
-
-import SparkExceptions.SparkException;
-
-public class FileException extends SparkException {
-    public FileException(String s) {
-        super(String.format("Encountered an issue while handling save-file: %s", s));
-    }
-}
