--- src/main/java/SparkExceptions/FileExceptions/ReadFileException.java
@@ -1,7 +1,0 @@
-package SparkExceptions.FileExceptions;
-
-public class ReadFileException extends FileException {
-    public ReadFileException() {
-        super("Save-File could not be read!");
-    }
-}
