--- src/main/java/SparkExceptions/FileExceptions/TaskDecodingException.java
@@ -1,9 +1,0 @@
-package SparkExceptions.FileExceptions;
-
-import SparkExceptions.SparkException;
-
-public class TaskDecodingException extends SparkException {
-    public TaskDecodingException() {
-        super("Could not decode a Task from the TaskFile! (was it corrupted?)");
-    }
-}
