--- src/main/java/SparkExceptions/FileExceptions/WriteFileException.java
@@ -1,7 +1,0 @@
-package SparkExceptions.FileExceptions;
-
-public class WriteFileException extends FileException {
-    public WriteFileException() {
-        super("Save-File could not be written to!");
-    }
-}
