--- src/main/java/SparkExceptions/FormatExceptions/FormatException.java
@@ -1,9 +1,0 @@
-package SparkExceptions.FormatExceptions;
-
-import SparkExceptions.SparkException;
-
-public class FormatException extends SparkException {
-    public FormatException(String errorMessage) {
-        super(errorMessage);
-    }
-}
