--- src/main/java/SparkExceptions/SparkException.java
@@ -1,7 +1,0 @@
-package SparkExceptions;
-
-public class SparkException extends Exception {
-    public SparkException(String errorMessage) {
-        super(String.format("OOPS!!!\n%s",errorMessage));
-    }
-}
