--- src/main/java/SparkExceptions/TaskModificationExceptions/InvalidTaskIdException.java
@@ -1,9 +1,0 @@
-package SparkExceptions.TaskModificationExceptions;
-
-import SparkExceptions.SparkException;
-
-public class InvalidTaskIdException extends SparkException {
-    public InvalidTaskIdException(String invalidInput) {
-        super(String.format("Your input:\n  %s\ndoes not seem like an integer. Task numbers are integers!", invalidInput));
-    }
-}
