--- src/main/java/SparkExceptions/TaskModificationExceptions/TaskAlreadyUnMarked.java
@@ -1,10 +1,0 @@
-package SparkExceptions.TaskModificationExceptions;
-
-import SparkExceptions.SparkException;
-import Tasks.TaskTypes.Task;
-
-public class TaskAlreadyUnMarked extends SparkException {
-    public TaskAlreadyUnMarked(Task t) {
-        super(String.format("   %s\nis already unmarked!", t));
-    }
-}
