--- src/main/java/SparkExceptions/TaskModificationExceptions/TaskNotFoundException.java
@@ -1,9 +1,0 @@
-package SparkExceptions.TaskModificationExceptions;
-
-import SparkExceptions.SparkException;
-
-public class TaskNotFoundException extends SparkException {
-    public TaskNotFoundException() {
-        super("I couldn't find the Task that you wanted :(");
-    }
-}
