+++ src/main/java/spark/commands/commandtypes/AddEventCommand.java
@@ -1,0 +1,28 @@
+package spark.commands.commandtypes;
+
+import spark.exceptions.SparkException;
+import spark.storage.Storage;
+import spark.tasks.TaskList;
+import spark.Ui;
+
+public class AddEventCommand extends Command {
+    private String[] tokens;
+
+    public AddEventCommand(String[] tokens) {
+        this.tokens = tokens;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        try {
+            tasks.addEvent(tokens);
+        } catch (SparkException e) {
+            ui.printException(e);
+        }
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
