+++ src/main/java/spark/commands/commandtypes/AddToDoCommand.java
@@ -1,0 +1,30 @@
+package spark.commands.commandtypes;
+
+import spark.exceptions.SparkException;
+import spark.storage.Storage;
+import spark.tasks.TaskList;
+import spark.Ui;
+
+public class AddToDoCommand extends Command {
+    private String[] tokens;
+
+    public AddToDoCommand(String[] tokens) {
+        this.tokens = tokens;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        try {
+            tasks.addToDo(tokens);
+        } catch (SparkException e) {
+            ui.printException(e);
+        }
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
+
+
