+++ src/main/java/spark/commands/commandtypes/MarkCommand.java
@@ -1,0 +1,28 @@
+package spark.commands.commandtypes;
+
+import spark.Ui;
+import spark.exceptions.SparkException;
+import spark.storage.Storage;
+import spark.tasks.TaskList;
+
+public class MarkCommand extends Command {
+    private String[] tokens;
+
+    public MarkCommand(String[] tokens) {
+        this.tokens = tokens;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        try {
+            tasks.markTask(tokens);
+        } catch (SparkException e) {
+            ui.printException(e);
+        }
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
