+++ src/main/java/spark/exceptions/SparkException.java
@@ -1,0 +1,7 @@
+package spark.exceptions;
+
+public class SparkException extends Exception {
+    public SparkException(String errorMessage) {
+        super(String.format("OOPS!!!\n%s",errorMessage));
+    }
+}
