+++ src/main/java/spark/exceptions/fileexceptions/ReadFileException.java
@@ -1,0 +1,7 @@
+package spark.exceptions.fileexceptions;
+
+public class ReadFileException extends FileException {
+    public ReadFileException() {
+        super("Save-File could not be read!");
+    }
+}
