+++ src/main/java/spark/exceptions/fileexceptions/WriteFileException.java
@@ -1,0 +1,7 @@
+package spark.exceptions.fileexceptions;
+
+public class WriteFileException extends FileException {
+    public WriteFileException() {
+        super("Save-File could not be written to!");
+    }
+}
