+++ src/main/java/spark/exceptions/formatexceptions/FormatException.java
@@ -1,0 +1,9 @@
+package spark.exceptions.formatexceptions;
+
+import spark.exceptions.SparkException;
+
+public class FormatException extends SparkException {
+    public FormatException(String errorMessage) {
+        super(errorMessage);
+    }
+}
