+++ src/main/java/spark/exceptions/taskmodificationexceptions/InvalidTaskIdException.java
@@ -1,0 +1,9 @@
+package spark.exceptions.taskmodificationexceptions;
+
+import spark.exceptions.SparkException;
+
+public class InvalidTaskIdException extends SparkException {
+    public InvalidTaskIdException(String invalidInput) {
+        super(String.format("Your input:\n  %s\ndoes not seem like an integer. Task numbers are integers!", invalidInput));
+    }
+}
