+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ----------------------------------------------------------------------
-Hello! I'm Spark
+Hello! I'm spark.spark
 What can I do for you?
 ----------------------------------------------------------------------
 ----------------------------------------------------------------------
